Reduce DriverManager.getDriver calls
